[#295] refactor: 반응형 위해 컴포넌트 분리
